Added the devices that matched the filter as a new field in the control units
This was done so we don't have to first make query to see which control
units matched the filter and then make a seperate query to find which
devices that matched the filter. Instead now we can find all the control
units that matches the filter and in each control unit we can find the
matching devices.