Send the id of the component to the 'view component' template instead of component itself
If we send the component object to the 'view component' template then we
lose the reactive updates of the component. So instead we send the id of
the component and the 'view component' template will fetch and display
the component and any updates to the component will be displayed in the
template automatically.