Added 'triggerObserver' library.
This library is used for sending specific actions to control units when
the trigger(s) is fulfilled. Example of a trigger is when a device
enters a certain area or when a property of a component changes to a
certain value. There can exist one or more triggers that must be
fulfilled before the specified action(s) is sent to the control unit.